+++ @types/function/index.d.ts
@@ -32,6 +32,16 @@ interface Window {
     readonly formatAsDisplayedMessage: typeof formatAsDisplayedMessage;
     readonly retrieveDisplayedMessage: typeof retrieveDisplayedMessage;
 
+    // import_raw
+    readonly importRawCharacter: typeof importRawCharacter;
+    readonly importRawPreset: typeof importRawPreset;
+    readonly importRawWorldbook: typeof importRawWorldbook;
+    readonly importRawTavernRegex: typeof importRawTavernRegex;
+
+    // inject
+    readonly injectPrompts: typeof injectPrompts;
+    readonly uninjectPrompts: typeof uninjectPrompts;
+
     // generate
     readonly builtin_prompt_default_order: typeof builtin_prompt_default_order;
     readonly generate: typeof generate;